# static/swagger.yaml
openapi: 3.0.0
info:
  title: Stellar Test Anchor
  description: An implementation of a SEP compliant Stellar Anchor
  version: 1.0.0
paths:
  /v1/authorizations:
    get:
      summary: Request A Challenge
      description: Requests a challenge transaction for the client to sign. Once signed, the challenge transaction is to be resubmitted back to the api.
      operationId: requestAChallenge
      tags:
        - Authorization
      parameters:
        - name: account
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Challenge Transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeTransaction'
          links:
            resubmit:
              $ref: '#/components/links/ChallengeTransactionResubmit'
    post:
      summary: Authenticate A Challenge
      description: Authenticates the user with a signed JWT token
      operationId: authenticate
      tags:
        - Authorization
      parameters:
        - name: account
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeTransaction'
      responses:
        '200':
          description: JWT Authentication Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
  /v1/deposit:
    get:
      summary: Get Account Deposit Details
      description: Gets the details for depositing currency into an account
      operationId: deposit
      tags:
        - Account
      security:
        - bearerAuth: []
      parameters:
        - name: account
          in: query
          required: true
          schema:
            type: string
        - name: asset_code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The deposit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDepositDetails'
components:
  securitySchemes:
    bearerAuth:
      type: middleware
      scheme: bearer
      bearerFormat: JWT
  links:
    ChallengeTransactionResubmit:
      operationId: authenticate
      parameters:
        transaction: $response.body#/transaction
  schemas:
    AccountDepositDetails:
      type: object
      properties:
        how:
          type: string
          example: 0x2C4c80B97de52AD03f1759cCC41747C76934ee3D
    ChallengeTransaction:
      type: object
      properties:
        transaction:
          type: string
          example: AAAAAP3z5PruosqouovJ83RCC6WM1tYr8e4hZHHFC1CDBcXKAAAAZAAAAAAAAAAAAAAAAQAAAABdMUqUAAAAAF0xS8AAAAAAAAAAAQAAAAEAAAAAdV2iRRpbBLR3/Xjf+xxG0JLMVClrKtmUox/sjbp7bbsAAAAKAAAAFlN0ZWxsYXIgRkkgQW5jaG9yIGF1dGgAAAAAAAEAAABAYXk2WEE3LVRYejJiTFFQZEYyT05raXRRZ2RfNEZ0dk1aS2Zwc1BLd1pyZU96LUwwNHZkSk5OLURBU2pFZWx5MAAAAAAAAAABgwXFygAAAEDU15m0XSyjln1uObo+MPu5gFwZOVg+dm1iS89KPYNckzK9O//5rJqEDl2u8YRKKRrcNaVnnxF5iJWKLnQA8mcP
      required:
        - transaction
    AuthToken:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1NjM1OTc5NzYsImlhdCI6MTU2MzUxMTU3NiwiaXNzIjoiIiwianRpIjoiMmNhZDNiMWE1Yjk3MzllYzZiYzI5M2Y4ZTJiNjFiYmU3NjRhY2RkOTBlMTQ1MTQ1MGI2Nzg2OGIxNDIwYmNmMCIsInN1YiI6IkdCMlYzSVNGREpOUUpORFg3VjRONzZZNEkzSUpGVENVRkZWU1ZXTVVVTVA2WkROMlBOVzNXWjY0In0.TOgt5WkwNFrx0wZU23kA11to59pxhFR8V6mB22mUI5I
      required:
        - token